Visual Styling - Particles complete
public ParticleSystem

Instantiating a Particle System
for an effect that happens only once (like being hit or grabbing a health collectible), you can store a reference to the Prefab in a public variable and call Instantiate when it should happen. If the Particle System is not Looping and Stop Action is set to Destroy, it will play and then be destroyed.