Visual Styling - User Interface - Head-Up Display
as you’ve seen before, static members are shared across all instances of that scripts, so it is the same value in all the GameObjects with that script attached to them.
If you have two Health bars in your Scene, the second one will store itself in the static member too, replacing the first one. So UIHealthBar.instance would always return the second one, never the first one. This is why that particular setup is called a Singleton, because only one object of that type can exist. This is exactly what you want here: to have only one Health bar.